+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/ContributionHintFirGenerator.kt
@@ -21,10 +21,12 @@ import dev.zacsweers.metro.compiler.ir.transformers.HintGenerator
 import dev.zacsweers.metro.compiler.mapNotNullToSet
 import dev.zacsweers.metro.compiler.scopeHintFunctionName
 import dev.zacsweers.metro.compiler.symbols.Symbols
+import org.jetbrains.kotlin.descriptors.Visibilities
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.caches.FirCache
 import org.jetbrains.kotlin.fir.caches.firCachesFactory
 import org.jetbrains.kotlin.fir.declarations.toAnnotationClassIdSafe
+import org.jetbrains.kotlin.fir.declarations.utils.visibility
 import org.jetbrains.kotlin.fir.extensions.ExperimentalTopLevelDeclarationsGenerationApi
 import org.jetbrains.kotlin.fir.extensions.FirDeclarationGenerationExtension
 import org.jetbrains.kotlin.fir.extensions.FirDeclarationPredicateRegistrar
@@ -58,7 +60,10 @@ internal class ContributionHintFirGenerator(
         session.predicates.contributesAnnotationPredicate
       )
 
-    return (injectedClasses + contributedClasses).filterIsInstance<FirClassSymbol<*>>().distinct()
+    return (injectedClasses + contributedClasses)
+      .filterIsInstance<FirClassSymbol<*>>()
+      .filterNot { it.visibility == Visibilities.Private }
+      .distinct()
   }
 
   private val typeResolverFactory by lazy { MetroFirTypeResolver.Factory(session) }